refactor: clipper moved to sep component + moved up one directory + added story for it
